add documentation on communication
